delete fun preview

+++ app/src/main/java/com/sanaa/tudee_assistant/presentation/screen/tasks/TasksScreen.kt
@@ -34,33 +34,26 @@ import androidx.compose.ui.draw.rotate
 import androidx.compose.ui.platform.LocalLayoutDirection
 import androidx.compose.ui.res.painterResource
 import androidx.compose.ui.res.stringResource
-import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.LayoutDirection
 import androidx.compose.ui.unit.dp
 import com.sanaa.tudee_assistant.R
 import com.sanaa.tudee_assistant.presentation.composable.CustomDatePickerDialog
-import com.sanaa.tudee_assistant.presentation.composable.bottomSheet.DeleteComponent
 import com.sanaa.tudee_assistant.presentation.composable.TaskStatusTabs
+import com.sanaa.tudee_assistant.presentation.composable.bottomSheet.DeleteComponent
+import com.sanaa.tudee_assistant.presentation.composable.bottomSheet.task.AddEditTaskScreen
+import com.sanaa.tudee_assistant.presentation.composable.bottomSheet.task.TaskDetailsComponent
 import com.sanaa.tudee_assistant.presentation.designSystem.component.DayItem
 import com.sanaa.tudee_assistant.presentation.designSystem.component.TudeeSnackBar
 import com.sanaa.tudee_assistant.presentation.designSystem.component.button.FloatingActionButton
 import com.sanaa.tudee_assistant.presentation.designSystem.theme.Theme
-import com.sanaa.tudee_assistant.presentation.model.TaskUiStatus
 import com.sanaa.tudee_assistant.presentation.route.TasksScreenRoute
-import com.sanaa.tudee_assistant.presentation.composable.bottomSheet.task.AddEditTaskScreen
-import com.sanaa.tudee_assistant.presentation.composable.bottomSheet.task.TaskDetailsComponent
 import com.sanaa.tudee_assistant.presentation.state.TaskUiState
-import com.sanaa.tudee_assistant.presentation.utils.DataProvider
 import com.sanaa.tudee_assistant.presentation.utils.DateFormater
 import com.sanaa.tudee_assistant.presentation.utils.DateFormater.getShortMonthName
 import kotlinx.coroutines.launch
-import kotlinx.datetime.Clock
 import kotlinx.datetime.DateTimeUnit
-import kotlinx.datetime.LocalDate
-import kotlinx.datetime.TimeZone
 import kotlinx.datetime.minus
 import kotlinx.datetime.plus
-import kotlinx.datetime.toLocalDateTime
 import org.koin.androidx.compose.koinViewModel
 
 
@@ -97,18 +90,15 @@ fun TasksScreenContent(
     var taskToEdit by remember { mutableStateOf<TaskUiState?>(null) }
     var showDialog by remember { mutableStateOf(false) }
     var showAddTaskBottomSheet by remember { mutableStateOf(false) }
-    var daysInMonth by
-    remember {
+    var daysInMonth by remember {
         mutableStateOf(
             DateFormater.getLocalDatesInMonth(
-                state.selectedDate.year,
-                state.selectedDate.monthNumber
+                state.selectedDate.year, state.selectedDate.monthNumber
             )
         )
     }
     Box(
-        modifier = modifier
-            .fillMaxSize()
+        modifier = modifier.fillMaxSize()
     ) {
         Column(
             modifier = modifier
@@ -136,13 +126,10 @@ fun TasksScreenContent(
                         .clickable {
                             val nextMonth = state.selectedDate.minus(1, DateTimeUnit.MONTH)
                             interactionListener.onDateSelected(nextMonth)
-                            daysInMonth =
-                                (DateFormater.getLocalDatesInMonth(
-                                    nextMonth.year,
-                                    nextMonth.monthNumber
-                                ))
-                        },
-                    contentAlignment = Alignment.Center
+                            daysInMonth = (DateFormater.getLocalDatesInMonth(
+                                nextMonth.year, nextMonth.monthNumber
+                            ))
+                        }, contentAlignment = Alignment.Center
                 ) {
                     Icon(
                         painter = painterResource(R.drawable.icon_left_arrow),
@@ -170,8 +157,7 @@ fun TasksScreenContent(
                             .rotate(if (LocalLayoutDirection.current == LayoutDirection.Rtl) 90f else -90f)
                             .clickable {
                                 showDialog = true
-                            }
-                    )
+                            })
                 }
 
                 Box(
@@ -182,13 +168,10 @@ fun TasksScreenContent(
                         .clickable {
                             val previousMonth = state.selectedDate.plus(1, DateTimeUnit.MONTH)
                             interactionListener.onDateSelected(previousMonth)
-                            daysInMonth =
-                                (DateFormater.getLocalDatesInMonth(
-                                    previousMonth.year,
-                                    previousMonth.monthNumber
-                                ))
-                        },
-                    contentAlignment = Alignment.Center
+                            daysInMonth = (DateFormater.getLocalDatesInMonth(
+                                previousMonth.year, previousMonth.monthNumber
+                            ))
+                        }, contentAlignment = Alignment.Center
                 ) {
                     Icon(
                         painter = painterResource(R.drawable.icon_left_arrow),
@@ -219,48 +202,45 @@ fun TasksScreenContent(
             ) {
                 itemsIndexed(daysInMonth) { index, date ->
                     DayItem(
-                        dayDate = date,
-                        isSelected = date == state.selectedDate,
-                        onClick = {
+                        dayDate = date, isSelected = date == state.selectedDate, onClick = {
                             interactionListener.onDateSelected(date)
-                        }
-                    )
+                        })
                 }
             }
 
             if (showDialog) {
                 CustomDatePickerDialog(
                     onDateSelected = { selectedDateMillis: Long? ->
-                        selectedDateMillis?.let {
-                            val date = DateFormater.formatLongToDate(selectedDateMillis)
-                            interactionListener.onDateSelected(date)
-                            daysInMonth =
-                                (DateFormater.getLocalDatesInMonth(date.year, date.monthNumber))
-                        }
-                    },
-                    onDismiss = { showDialog = false },
-                    initialSelectedDate = state.selectedDate
+                    selectedDateMillis?.let {
+                        val date = DateFormater.formatLongToDate(selectedDateMillis)
+                        interactionListener.onDateSelected(date)
+                        daysInMonth =
+                            (DateFormater.getLocalDatesInMonth(date.year, date.monthNumber))
+                    }
+                }, onDismiss = { showDialog = false }, initialSelectedDate = state.selectedDate
                 )
             }
-            TaskStatusTabs(state, interactionListener::onTaskSwipeToDelete, interactionListener::onTaskClicked)
-            if (state.selectedTask != null && state.showTaskDetailsBottomSheet)
-                TaskDetailsComponent(
-                    task = state.selectedTask,
-                    onDismiss = { interactionListener.onDismissTaskDetails(false) },
-                    onEditClick = { task ->
-                        interactionListener.onDismissTaskDetails(false)
-                        taskToEdit = task
-                        showEditTaskBottomSheet = true
-                    },
-                    onMoveToClicked = interactionListener::onMoveTaskToAnotherStatus,
-                    modifier = Modifier.padding(horizontal = 16.dp)
-                )
-            if (state.selectedTask != null && state.showDeleteTaskBottomSheet)
-                DeleteComponent(
-                    onDismiss = interactionListener::onDeleteDismiss,
-                    onDeleteClicked = interactionListener::onDeleteTask,
-                    title = stringResource(R.string.delete_task_title),
-                )
+            TaskStatusTabs(
+                state,
+                interactionListener::onTaskSwipeToDelete,
+                interactionListener::onTaskClicked
+            )
+            if (state.selectedTask != null && state.showTaskDetailsBottomSheet) TaskDetailsComponent(
+                task = state.selectedTask,
+                onDismiss = { interactionListener.onDismissTaskDetails(false) },
+                onEditClick = { task ->
+                    interactionListener.onDismissTaskDetails(false)
+                    taskToEdit = task
+                    showEditTaskBottomSheet = true
+                },
+                onMoveToClicked = interactionListener::onMoveTaskToAnotherStatus,
+                modifier = Modifier.padding(horizontal = 16.dp)
+            )
+            if (state.selectedTask != null && state.showDeleteTaskBottomSheet) DeleteComponent(
+                onDismiss = interactionListener::onDeleteDismiss,
+                onDeleteClicked = interactionListener::onDeleteTask,
+                title = stringResource(R.string.delete_task_title),
+            )
             if (showAddTaskBottomSheet) {
                 AddEditTaskScreen(
                     isEditMode = false,
@@ -270,49 +250,38 @@ fun TasksScreenContent(
                         showAddTaskBottomSheet = false
                         coroutineScope.launch {
                             snackBarHostState.showSnackbar(
-                                message = "Task added successfully",
-                                withDismissAction = true
+                                message = "Task added successfully", withDismissAction = true
                             )
                         }
                     },
                     onError = { errorMessage ->
                         coroutineScope.launch {
                             snackBarHostState.showSnackbar(
-                                message = errorMessage,
-                                withDismissAction = true
+                                message = errorMessage, withDismissAction = true
                             )
                         }
-                    }
-                )
+                    })
             }
 
             if (showEditTaskBottomSheet && taskToEdit != null) {
-                AddEditTaskScreen(
-                    isEditMode = true,
-                    taskToEdit = taskToEdit,
-                    onDismiss = {
-                        showEditTaskBottomSheet = false
-                        taskToEdit = null
-                    },
-                    onSuccess = {
-                        showEditTaskBottomSheet = false
-                        taskToEdit = null
-                        coroutineScope.launch {
-                            snackBarHostState.showSnackbar(
-                                message = "Task updated successfully",
-                                withDismissAction = true
-                            )
-                        }
-                    },
-                    onError = { errorMessage ->
-                        coroutineScope.launch {
-                            snackBarHostState.showSnackbar(
-                                message = errorMessage,
-                                withDismissAction = true
-                            )
-                        }
+                AddEditTaskScreen(isEditMode = true, taskToEdit = taskToEdit, onDismiss = {
+                    showEditTaskBottomSheet = false
+                    taskToEdit = null
+                }, onSuccess = {
+                    showEditTaskBottomSheet = false
+                    taskToEdit = null
+                    coroutineScope.launch {
+                        snackBarHostState.showSnackbar(
+                            message = "Task updated successfully", withDismissAction = true
+                        )
                     }
-                )
+                }, onError = { errorMessage ->
+                    coroutineScope.launch {
+                        snackBarHostState.showSnackbar(
+                            message = errorMessage, withDismissAction = true
+                        )
+                    }
+                })
             }
         }
 
@@ -330,7 +299,7 @@ fun TasksScreenContent(
             iconRes = R.drawable.note_add
         )
 
-        var successMessageText : String? = null
+        var successMessageText: String? = null
         state.successMessageStringId?.let {
             successMessageText = stringResource(state.successMessageStringId)
         }
@@ -341,7 +310,7 @@ fun TasksScreenContent(
                 interactionListener.onShowSnackbar()
             }
         }
-        var errorMessageText : String? = null
+        var errorMessageText: String? = null
         state.errorMessageStringId?.let {
             successMessageText = stringResource(state.errorMessageStringId)
         }
@@ -353,58 +322,7 @@ fun TasksScreenContent(
         }
 
         TudeeSnackBar(
-            snackBarHostState = snackBarHostState,
-            isError = state.errorMessageStringId != null
+            snackBarHostState = snackBarHostState, isError = state.errorMessageStringId != null
         )
     }
 }
-
-//@Preview(showBackground = true)
-//@Composable
-//private fun TasksScreenPreview() {
-//    var tasksState by remember {
-//        mutableStateOf(
-//            TasksScreenUiState(
-//                currentDateTasks = DataProvider.getTasksSample(),
-//                selectedStatusTab = TaskUiStatus.DONE,
-//                selectedDate = Clock.System.now().toLocalDateTime(TimeZone.UTC).date,
-//            )
-//        )
-//    }
-//
-//    val previewInteraction = object : TaskInteractionListener {
-//        override fun onDateSelected(date: LocalDate) {
-//            tasksState = tasksState.copy(
-//                selectedDate = date,
-//                currentDateTasks = DataProvider.getTasksSample() // Simulate fetch
-//            )
-//        }
-//
-//        override fun onTaskSwiped(task: TaskUiState) {
-//            tasksState = tasksState.copy(
-//                currentDateTasks = tasksState.currentDateTasks.filterNot { it == task }
-//            )
-//        }
-//
-//        override fun onTaskClicked(task: TaskUiState) {
-//            tasksState = tasksState.copy(
-//                selectedTask = task,
-//                showTaskDetailsBottomSheet = true
-//            )
-//        }
-//
-//        override fun onDismissTaskDetailsDialog() {
-//            tasksState = tasksState.copy(showTaskDetailsBottomSheet = false)
-//        }
-//
-//        override fun onMoveTaskToAnotherStatus() = Unit
-//        override fun onTaskDeleted() = Unit
-//        override fun onTaskDeletedDismissed() = Unit
-//        override fun onSnackbarShown() = Unit
-//    }
-//
-//    TasksScreenContent(
-//        state = tasksState,
-//        interactionListener = previewInteraction
-//    )
-//}
